requirements.txt
wxPython
feedparser
requests
beautifulsoup4
yt-dlp
python-dateutil
mutagen
pyinstaller
